dale loco commit

+++ src/main/java/uniandes/isis2304/parranderos/negocio/VOBrindan.java
@@ -1,6 +1,8 @@
 package uniandes.isis2304.parranderos.negocio;
 
+import java.math.BigDecimal;
+
 public interface VOBrindan {
-    public long getIdHabitacion();
-    public long getIdServicioH();
+    public BigDecimal getIdHabitacion();
+    public BigDecimal getIdServicioH();
 }
