+++ src/main/java/uniandes/isis2304/parranderos/negocio/VOCliente.java
@@ -1,7 +1,9 @@
 package uniandes.isis2304.parranderos.negocio;
 
+import java.math.BigDecimal;
+
 public interface VOCliente {
-    public Long getId();
+    public BigDecimal getId();
     public String getNombre();
     public String getCorreo();
     public String getContrasenia();
