+++ src/main/java/uniandes/isis2304/parranderos/negocio/VOPropietarioInmueble.java
@@ -1,7 +1,9 @@
 package uniandes.isis2304.parranderos.negocio;
 
+import java.math.BigDecimal;
+
 public interface VOPropietarioInmueble {
-    public long getIdPi();
+    public BigDecimal getIdPi();
     public String getNombre();
     public String getVinculo();
     public String getTipo();
