+++ src/main/java/uniandes/isis2304/parranderos/negocio/VOSeguro.java
@@ -1,13 +1,14 @@
 package uniandes.isis2304.parranderos.negocio;
+import java.math.BigDecimal;
 import java.sql.Timestamp;
 
 public interface VOSeguro {
 
-    public Long getId();
+    public BigDecimal getId();
 
     public String getDescripcion();
 
-    public Long getIdInmueble();
+    public BigDecimal getIdInmueble();
 
     public Timestamp getFechaVence();
